feat: protects CargoController

+++ mobiauto-backend/src/main/java/com/mobiauto/backend/application/services/CargoService.java
@@ -1,34 +1,51 @@
 package com.mobiauto.backend.application.services;
 
+import com.mobiauto.backend.application.utils.AuthorizationUtils;
+import com.mobiauto.backend.domain.exceptions.Auth.UnauthorizedException;
 import com.mobiauto.backend.domain.exceptions.Cargo.CargoNotFoundException;
 import com.mobiauto.backend.domain.models.Cargo;
 import com.mobiauto.backend.domain.repositories.CargoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class CargoService {
 
     private final CargoRepository cargoRepository;
+    private final AuthorizationUtils authorizationUtils;
 
     @Autowired
-    public CargoService(CargoRepository cargoRepository) {
+    public CargoService(CargoRepository cargoRepository, AuthorizationUtils authorizationUtils) {
         this.cargoRepository = cargoRepository;
+        this.authorizationUtils = authorizationUtils;
     }
 
-    @PreAuthorize("hasRole('ADMIN')")
     public Cargo findById(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!authorizationUtils.isSuperuser(authentication)) {
+            throw new UnauthorizedException();
+        }
         return cargoRepository.findById(id).orElseThrow(CargoNotFoundException::new);
     }
 
-    @PreAuthorize("hasRole('ADMIN')")
+    @Transactional
     public Cargo save(Cargo cargo) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!authorizationUtils.isSuperuser(authentication)) {
+            throw new UnauthorizedException();
+        }
         return cargoRepository.save(cargo);
     }
 
-    @PreAuthorize("hasRole('ADMIN')")
+    @Transactional
     public void delete(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!authorizationUtils.isSuperuser(authentication)) {
+            throw new UnauthorizedException();
+        }
         Cargo cargo = findById(id);
         cargoRepository.delete(cargo);
     }
